+++ lib/active_attr/typecasting/boolean_typecaster.rb
@@ -17,6 +17,13 @@ module ActiveAttr
       # @since 0.5.0
       FALSE_VALUES = ["n", "N", "no", "No", "NO", "false", "False", "FALSE", "off", "Off", "OFF", "f", "F"]
 
+      # Values which force a nil result for typecasting
+      #
+      # These values are based on the behavior of ActiveRecord
+      #
+      # @since 0.14.0
+      NIL_VALUES = ["", nil]
+
       # Typecasts an object to true or false
       #
       # Similar to ActiveRecord, when the attribute is a zero value or
@@ -34,6 +41,7 @@ module ActiveAttr
       def call(value)
         case value
         when *FALSE_VALUES then false
+        when *NIL_VALUES then nil
         when Numeric, /\A[-+]?(0+\.?0*|0*\.?0+)\z/ then !value.to_f.zero?
         else value.present?
         end