+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ActiveAttr 0.14.0 (Unreleased) #
 
 * Drop support for Ruby versions below 1.9.2
+* #139 Changed Typecasting::BooleanTypecaster to cast nil and empty
+  strings to false (Corin Langosch)
 
 # ActiveAttr 0.13.1 (April 25, 2019) #